AMI DATASET-2023
Acute myocardial infarction (AMI) is the main cause of death in developed and developing
countries. AMI is a serious medical problem that necessitates hospitalization and sometimes results in death.
Patients hospitalized in the emergency department (ED) should therefore receive an immediate diagnosis and
treatment. Many studies have been conducted on the prognosis of AMI with hemogram parameters. However,
no study has investigated potential hemogram parameters for the diagnosis of AMI using an interpretable
artificial intelligence-based clinical approach. The purpose of this research is to implement the principles
of explainable artificial intelligence (XAI) in the analysis of hematological predictors for AMI. In this
retrospective analysis, 477 (48.6%) patients with AMI and 504 (51.4%) healthy individuals were enrolled
and assessed in predicting AMI. Of the patients with AMI, 182 (38%) had an ST-segment elevation MI
(STEMI), and 295 (62%) had a non-ST-segment elevation MI (NSTEMI). Demographic and hematological
information of the patients was analyzed to determine AMI. The XAI approach combined with machine
learning approaches (Extreme Gradient Boosting, XGB; Adaptive Boosting, AB; Light Gradient Boosting
Machine, LGBM) was applied for the estimation of AMI and distinguishing subgroups of AMI (STEMI
and NSTEMI). The SHAP approach was used to explain the predictions intuitively. After selecting the
10 most important hematological parameters for AMI, the LGBM model achieved 83% and 74% accuracy for
prediction of AMI, and distinguishing subgroups of AMI (STEMI and NSTEMI), respectively. SHAP results
showed that neutrophil (NEU), white blood cell (WBC), platelet width of distribution (PDW), and basophil
(BA) were the most important for AMI prediction. Mean corpuscular volume (MCV), BA, monocytes (MO),
and lymphocytes (LY) were the most important hematological parameters that distinguish STEMI from
NSTEMI. The proposed model serves as a valuable tool for physicians, facilitating the diagnosis, treatment,
and follow-up of patients with AMI and distinguishing subgroups of AMI (STEMI and NSTEMI). Analyzing
readily accessible hemogram parameters empowers medical professionals to make informed decisions and
provide enhanced care to a wide range of individuals.